refactor(backend): extract source crud module
Move the source and chunk functions out of crud/__init__.py into
crud/source.py, following the split already done for comment, token,
metric and others. Consumers import the module directly; nothing is
re-exported.

Drops update_chunk, delete_chunk and get_chunk, which had no callers
anywhere in the repo, and the ChunkUpdate schema they were the only
users of. create_chunk stays -- ChunkingService calls it.
