+++ apps/backend/src/rhesis/backend/app/crud/__init__.py
@@ -844,154 +844,6 @@ def delete_status(
     return delete_item(db, models.Status, status_id, organization_id, user_id)
 
 
-# Source CRUD
-_SOURCE_RELATED_FIELDS = (
-    include(models.Source.source_type),
-    include(models.Source.user),
-)
-
-
-def get_source(
-    db: Session,
-    source_id: uuid.UUID,
-    organization_id: str = None,
-    user_id: str = None,
-) -> Optional[models.Source]:
-    """Get source.
-
-    Note: Content field is deferred and will not be loaded unless explicitly requested.
-    Use get_source_with_content() to load the content field.
-    Relationships (source_type, user) are loaded for display.
-    """
-    from rhesis.backend.app.utils.crud_utils import _check_and_raise_if_deleted
-
-    item = (
-        QueryBuilder(db, models.Source)
-        .with_deleted()
-        .with_related(*_SOURCE_RELATED_FIELDS)
-        .with_default_derived_field_loads()
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .filter_by_id(source_id)
-    )
-    return _check_and_raise_if_deleted(item, models.Source, source_id, False)
-
-
-def get_source_with_content(
-    db: Session,
-    source_id: uuid.UUID,
-    organization_id: str = None,
-    user_id: str = None,
-    include_deleted: bool = False,
-) -> Optional[models.Source]:
-    """Get source with content field explicitly loaded (a deferred column)."""
-    from sqlalchemy.orm import undefer
-
-    from rhesis.backend.app.utils.crud_utils import _check_and_raise_if_deleted
-
-    item = (
-        QueryBuilder(db, models.Source)
-        .with_deleted()
-        .with_related(*_SOURCE_RELATED_FIELDS)
-        .with_default_derived_field_loads()
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-    )
-    item.query = item.query.options(undefer(models.Source.content))
-    item = item.filter_by_id(source_id)
-    return _check_and_raise_if_deleted(item, models.Source, source_id, include_deleted)
-
-
-def get_sources(
-    db: Session,
-    skip: int = 0,
-    limit: int = 10,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Source]:
-    """Get sources.
-
-    Note: Content field is deferred and will not be loaded.
-    Use get_source_with_content() for individual sources that need content.
-    Relationships (source_type, user) are loaded for display.
-    """
-    return (
-        QueryBuilder(db, models.Source)
-        .with_related(*_SOURCE_RELATED_FIELDS)
-        .with_default_derived_field_loads()
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .with_odata_filter(filter)
-        .with_sorting(sort_by, sort_order)
-        .with_pagination(skip, limit)
-        .all()
-    )
-
-
-def create_source(
-    db: Session, source: schemas.SourceCreate, organization_id: str = None, user_id: str = None
-) -> models.Source:
-    """Create source."""
-    return create_item(db, models.Source, source, organization_id, user_id)
-
-
-def update_source(
-    db: Session,
-    source_id: uuid.UUID,
-    source: schemas.SourceUpdate,
-    organization_id: str = None,
-    user_id: str = None,
-) -> Optional[models.Source]:
-    """Update source."""
-    return update_item(db, models.Source, source_id, source, organization_id, user_id)
-
-
-def delete_source(
-    db: Session, source_id: uuid.UUID, organization_id: str, user_id: str
-) -> Optional[models.Source]:
-    """Delete source."""
-    return delete_item(db, models.Source, source_id, organization_id, user_id)
-
-
-def create_chunk(
-    db: Session, chunk: schemas.ChunkCreate, organization_id: str = None, user_id: str = None
-) -> models.Chunk:
-    """Create chunk."""
-    return create_item(db, models.Chunk, chunk, organization_id=organization_id, user_id=user_id)
-
-
-def update_chunk(
-    db: Session,
-    chunk_id: uuid.UUID,
-    chunk: schemas.ChunkUpdate,
-    organization_id: str = None,
-    user_id: str = None,
-) -> Optional[models.Chunk]:
-    """Update chunk."""
-    return update_item(
-        db, models.Chunk, chunk_id, chunk, organization_id=organization_id, user_id=user_id
-    )
-
-
-def delete_chunk(
-    db: Session, chunk_id: uuid.UUID, organization_id: str, user_id: str
-) -> Optional[models.Chunk]:
-    """Delete chunk."""
-    return delete_item(db, models.Chunk, chunk_id, organization_id=organization_id, user_id=user_id)
-
-
-def get_chunk(
-    db: Session, chunk_id: uuid.UUID, organization_id: str = None, user_id: str = None
-) -> Optional[models.Chunk]:
-    """Get chunk."""
-    return get_item_detail(
-        db, models.Chunk, chunk_id, organization_id=organization_id, user_id=user_id
-    )
-
-
 # Topic CRUD
 def get_topic(
     db: Session, topic_id: uuid.UUID, organization_id: str = None, user_id: str = None
